refactor: split channel-delivery-store.ts into focused modules
